+++ web/src/app/builds/BuildsBrowser.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { useEffect, useRef, useState, useTransition } from "react";
+import { humanSize } from "@/lib/format";
 import { usePathname, useRouter } from "next/navigation";
 import RowLink from "./RowLink";
 import MassApply from "@/components/MassApply";
@@ -8,18 +9,6 @@ import Select from "@/components/Select";
 import { useModerator } from "@/components/useModerator";
 import { buildHref } from "@/lib/slug";
 import type { BuildListItem, BuildSortKey } from "@/lib/queries";
-
-function humanSize(bytes?: number): string {
-  if (bytes == null) return "—";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  let v = Number(bytes);
-  let i = 0;
-  while (v >= 1024 && i < units.length - 1) {
-    v /= 1024;
-    i++;
-  }
-  return i === 0 ? `${v} B` : `${v.toFixed(1)} ${units[i]}`;
-}
 
 function formatDate(date: string | null): string {
   if (!date) return "—";
