+++ web/src/app/builds/[buildId]/page.tsx
@@ -1,4 +1,5 @@
 import { Fragment } from "react";
+import { humanSize } from "@/lib/format";
 import type { Metadata } from "next";
 import Link from "next/link";
 import { notFound, permanentRedirect } from "next/navigation";
@@ -33,18 +34,6 @@ export const runtime = "nodejs";
 export const revalidate = 3600;
 export function generateStaticParams(): Array<{ buildId: string }> {
   return [];
-}
-
-function humanSize(bytes?: number): string {
-  if (bytes == null) return "—";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  let v = bytes;
-  let i = 0;
-  while (v >= 1024 && i < units.length - 1) {
-    v /= 1024;
-    i++;
-  }
-  return i === 0 ? `${bytes} B` : `${v.toFixed(1)} ${units[i]}`;
 }
 
 function titleize(k: string): string {
