+++ web/src/app/games/[slug]/GameBuilds.tsx
@@ -5,24 +5,13 @@
 // mis-filed build to another game, clear it, or move builds between lots).
 
 import { useState } from "react";
+import { humanSize } from "@/lib/format";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import MassApply from "@/components/MassApply";
 import { useModerator } from "@/components/useModerator";
 import { buildHref } from "@/lib/slug";
 import type { BuildListItem } from "@/lib/queries";
-
-function humanSize(bytes?: number): string {
-  if (bytes == null) return "—";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  let v = Number(bytes);
-  let i = 0;
-  while (v >= 1024 && i < units.length - 1) {
-    v /= 1024;
-    i++;
-  }
-  return i === 0 ? `${v} B` : `${v.toFixed(1)} ${units[i]}`;
-}
 
 export default function GameBuilds({ builds }: { builds: BuildListItem[] }) {
   const router = useRouter();
